Basic background scan functionality implemented
Background scans are now enabled. Only one scan (foreground or background) may run at a time, so starting a foreground scan while a background scan is in progress will result in the foreground scan waiting for the background scan to complete before running. Also implemented filter by "rarity". At this point, rarity is hardcoded but it will be made configurable at a future date.

There will be a potential issue if the number of accounts is too low and the scan frequency too high as the rate of scans being triggered cannot keep up with how fast each scan takes to complete.